requirements.txt
streamlit
pandas
numpy
plotly
pydeck
Pillow
openpyxl
    